# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API FocoMultimida",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/reserve": {
            "post": {
                "tags": [
                    "Reserve"
                ],
                "summary": "Create a new reservation",
                "description": "This endpoint allows the creation of a new reservation.",
                "operationId": "createReservation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "CheckIn",
                                    "CheckOut",
                                    "hotelCode",
                                    "roomCode",
                                    "dailyValue",
                                    "guestName",
                                    "guestLastName",
                                    "guestPhone"
                                ],
                                "properties": {
                                    "CheckIn": {
                                        "description": "Check-in date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-11-03"
                                    },
                                    "CheckOut": {
                                        "description": "Check-out date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-11-05"
                                    },
                                    "hotelCode": {
                                        "description": "Hotel code",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "roomCode": {
                                        "description": "Room code",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dailyValue": {
                                        "description": "Daily value of the room",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "guestName": {
                                        "description": "Guest name",
                                        "type": "string",
                                        "example": "Anthonius"
                                    },
                                    "guestLastName": {
                                        "description": "Guest last name",
                                        "type": "string",
                                        "example": "Souza"
                                    },
                                    "guestPhone": {
                                        "description": "Guest phone number",
                                        "type": "string",
                                        "example": "557799925474"
                                    },
                                    "paymentMethod": {
                                        "description": "Payment method",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "paymentValue": {
                                        "description": "Payment value",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status message",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Reservation data",
                                            "properties": {
                                                "id": {
                                                    "description": "Reservation ID",
                                                    "type": "integer"
                                                },
                                                "CheckIn": {
                                                    "description": "Check-in date",
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "CheckOut": {
                                                    "description": "Check-out date",
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "hotelCode": {
                                                    "description": "Hotel code",
                                                    "type": "integer"
                                                },
                                                "roomCode": {
                                                    "description": "Room code",
                                                    "type": "integer"
                                                },
                                                "dailyValue": {
                                                    "description": "Daily value of the room",
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "guestCode": {
                                                    "description": "Guest code",
                                                    "type": "integer"
                                                },
                                                "Total": {
                                                    "description": "Total value of the reservation",
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "isPayed": {
                                                    "description": "Payment status",
                                                    "type": "boolean"
                                                },
                                                "created_at": {
                                                    "description": "Creation date",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "description": "Update date",
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status message",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "Room not available for the selected date or The check-out date must be greater than the check-in date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}